Add helpful print functions and fix AllSubTreesToString
Functions to turn a polnode and slices of polnodes to a string are
added.

AllSubTreesToString is changed to not rely on calculatePosition, which
helps in the case of there being a bug with the pollard. There are
cases where the a subtree wouldn't print because it couldn't calculate
the position. Those bugs are now eliminated.