accumulator: Fix one bug with undoing empty roots
Undoing empty roots was done being incorrectly as the slice of deletions
were not being sorted and they weren't being added from greatest to the
lowest position. These are fixed in this commit.

However, there's still a bug with undoing empty roots as undoing empty
roots for additions needs more data then just the positions being
deleted and their hashes. For example:

The below pollard has 2 leaves but 8 numLeaves.

14
|---------------\
12              13
|-------\       |-------\

|---\   |---\   |---\   |---\

Undoing 1 addition may result in either one of these pollards:

|---------------\
--
|-------\       |-------\
--      --      --
|---\   |---\   |---\   |---\
--  --  --  --  --  --  06

|---------------\
--
|-------\       |-------\
--      --      10
|---\   |---\   |---\   |---\
--  --  --  --          --

|---------------\
12
|-------\       |-------\
--      --      --
|---\   |---\   |---\   |---\
--  --  --  --  --  --  --

There needs to be extra information saved to note which root was empty
and the code is still buggy because this information isn't provided.