accumulator_test: Remove TestPollardAddDel
TestPollardAddDel was using the simChain to generate random deletions.
This test would be better served as a fuzz test since we can now
automate the failures instead of having to manually test for them.

The test is removed in this commit and will be re-added as a fuzz test
in a later commit.