accumulator_test: Add fuzz tests for Modify
Previously removed TestPollardAddDel is now replaced with
FuzzPollardModifyChain. This provides an automated way to test for
failures that happened with simChain when using different durations and
random seeds.

FuzzModify is also added. This provides a fuzzing without the chain and
only goes through 2 total modifications. The pollard trees are also
smaller, making it easier to visualize when fixing a failure.