accumulator_test: Refactor TestUndo
TestUndo would only be able to test on initial pollards that never had a
deletion performed. This refactor allows testing of undos for pollards
that has had a deletion now.

Previously we can only have these pollards to start off of:

02
|---\
00  01

14
|---------------\
12              13
|-------\       |-------\
08      09      10      11
|---\   |---\   |---\   |---\
00  01  02  03  04  05  06  07

Now we can have these pollards to start off of:

02
|---\

14
|---------------\
12              13
|-------\       |-------\
                10      11
|---\   |---\   |---\   |---\
                        06  07