prove: Add proofAfterDeletion function
proofAfterDeletion function takes in a proof and returns a new proof
that can be used to calculate what the roots would be after the deletion
of the targets in the current proof has happened.

This function is useful in calculating the roots of the accumulator in a
purely functional way.