stump: Add new accumulator representation called stump
Stump is a minimum required data needed for validating the accumulator
proofs. The advantages after that it doesn't have any of the
complexities that pollard has and is much simpler code-wise.

It seems that a user can keep track of proofs separately while using
stump to verify and update the accumulator state. Worth looking at as
it'll be much simpler than pollard.