stump_test: Add tests for stump
FuzzStump and FuzzStumpChain are added. Like the existing tests such as
FuzzPollardModify and FuzzPollardModifyChain, the basic idea is the
same.

FuzzStump does 2 modifications on the stump and is more beneficial
during debugging as you can limit the amount of leaves, making it easier
to visualize.

FuzzStumpChain uses a simulation blockchain to tests for
errors. It's much harder to visualize but having more modifications
allows for better testing of Stump.