prove: Reverse the rootCandidates in calculateRoots
The roots are organized from the greatest position to the smallest in
pollard. However in calculateRoots, since we're hashing from the
smallest position to the greatest and appending calculated roots along
the way, the roots are in smallest to greatest order. Reversing makes
the roots in the same order as how the roots are organized in pollard.