prove: Have one generic mergeSortedSlices
There were two mergeSortedSlices functions for two different types of
slices. We can have one function if we use generics.

The mergeSortedNodeAndPos wasn't used but I'm leaving making the
mergeSortedSlices function generic nonetheless as this library is going
through a lot of change and more flexiblility is worth it at the moment.