prove: Add GetMissingPositions()
GetMissingPositions() provides functionality for calculating the
positions you're missing to prove additional targets. This is useful
for mempool transactions relay as this allows for the peer to only
ask for certain un-cached positions instead of the entire proof.

There isn't code for constructing a proof from the missing positions
yet.