prove: Copy instead of returning slice in mergeSortedSlicesFunc()
mergeSortedSlicesFunc() would return the hashes that were passed in from
the caller when either slice a or slice b was empty. The expected result
from mergeSortedSlicesFunc() is to return a new slice so that other
slices are not mutated. But the existing logic made it so that the
returned slice was the passed in slice, resulting in bugs.

The new logic always allocates a new copy.