prove: Refactor proofAfterDeletion into proofAfterPartialDeletion
proofAfterDeletion only allowed for deleting all the targets in a proof.
However, if we want to modify proof for a wallet, we need the ability to
only delete select targets in a proof. ProofAfterPartialDeletion allows
for that and we keep proofAfterDeletion as a wrapper around
proofAfterPartialDeletion, still allowing callers to call without the
delHashes for the proof when deleting all the targets.