Prove: Modify calculateRoots() to support deletions
The changes made to calculateRoots() now enables support for calculating
the roots that happen after a deletion. This replaces the functionality
that proofAfterDeletion() along with the old calculateRoots() provided
for stump deletion. The deletion is now just M*log(N) where N = number of
leaves and M = number of targets.