utils: Simplify treeRows()
treeRows() is changed to a much simpler algorithm that only depends on
counting the number of bits used to represent the number of leaves.

Since only numLeaves of 2^n should decrement in the number of bits used,
the function should return the correct amount of total rows except 0
(which is handled separately).

This changed is based off of https://github.com/utreexo/pytreexo/pull/3