prove: Add support for undoing cached proofs
Cached proof are meant for wallets to be able to keep track of their own
utxos. Wallets must be able to reorg back blocks which means that the
cached proofs need to support undoing.

The Undo function mimics that of the Bitcoin Core wallet where the
reorging of a block where a spend occured doesn't automatically re-add
back the spent utxo. That must be done separately.