utreexo: abstract out String() and SubTreeToString() to functions
String() and SubTreeToString() were methods on Pollard. However, they
can easily be turned into functions on an interface, making it possible
for any type to use these functions for debugging as long as they adhere
to the ToString interface. Both of the methods on Pollard now call each
of the functions instead.

This is to allow for the partial accumulator implemented with maps
(maputreexo) to also use these functions instead of having to copy the
methods.