utreexo/utils: Add translatePos
translatePos is useful for when translating a given position to a
different total rows. There are some implementations that benefit from
pre-allocating rows. For these implementations, a given position may
differ in value if it's not in row 0.

The below illustrates how the positions would differ in different
pre-allocated total rows. position 2 in totalrows=1 would be position 4
if totalrows=2.

02
|---\
00  01

|-------\
04
|---\   |---\
00  01