utreexo: Avoid potential bug with isRootPosition
The totalRows argument in isRootPosition depended on the totalRows
being the minimum possible for the numleaves. For example, it's possible
to have a numleaves of 1 but have a totalRows of 50.

However, for isRootPosition to work correctly, totalRows always needed
to equal treeRows(numLeaves). This meant that if you gave position=0,
numleaves=1, totalRows=50, isRootPosition would return false even though
it should return true.

Because of this, isRootPosition and isRootPositionOnRow is changed to
only take in the numleaves to avoid future bugs.