mappollard: add GetMissingPositions
For p2p communication where a peer explicitly asks for the hashes it
needs, there needs to be accumulator support returning what positions a
node is missing to be able to prove certain leaves.

GetMissingPositions provides that functionality by returning the
positions of the hashes it's missing in ascending order.