mappollard: support full mappollard for Undo
For Undo(), mappollard was assuming that there weren't nodes to fetch
from since not all of the leaves are cached. But since full caches
everything, we don't need this assumption. The changes here make it
possible to Undo with just the targets of the proof and not the full
proof.