mappollard: add view with remove and add methods
A view represents a part of the mappollard with all the roots. When
something is removed from the view, its key is stored in the removed
slice.

The view mainly serves as a buffer to avoid lookups from the main map
which is slow since its so big. However, it also has the added benefit
of making removes atomic in mappollard.