mappollard: use view for Modify
We switch over to a view based Modify which has the benefit of avoiding
reads from the large nodes map for bridgenodes and adds the benefit of
the accumulator now being atomic on Modifies. Before an error during a
modify would cause it to be left in an inconsistent state but it's no
longer the case now.